\name{bmms-package}
\alias{bmms-package}
\alias{bmms}
\docType{package}
\title{
\packageTitle{bmms}
}
\description{
\packageDescription{bmms}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bmms}
\packageIndices{bmms}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bmms}

Maintainer: \packageMaintainer{bmms}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
