% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmms-package.R
\docType{package}
\name{bmms}
\alias{bmms}
\alias{bmms-package}
\title{bmms}
\description{
Bayesian Modular & Multiscale Regression
}
\details{
https://arxiv.org/abs/1809.05935
}
\author{
Michele Peruzzi <peruzzi@gmail.com>
}
