% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bvs}
\alias{bvs}
\title{A simple Bayesian Variable Selection model using g-priors}
\usage{
bvs(y, X, prior, mcmc, g = -1, model_prior_par = 1, fixs = FALSE)
}
\arguments{
\item{y}{vector of responses}

\item{X}{design matrix}

\item{prior}{starting model (#name to be changed#)}

\item{mcmc}{number of Markov chain Monte Carlo iterations}

\item{g}{g-prior parameter}

\item{model_prior_par}{For model M, p(M) is prop. to exp(k * p) where p is the 
number of included variables, and k is model_prior_par}

\item{fixs}{Fix the regression variance to 1?}
}
\description{
A simple Bayesian Variable Selection model using g-priors
}
