% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cube_mean}
\alias{cube_mean}
\title{Compute the mean along a cube dimension}
\usage{
cube_mean(X, dim)
}
\arguments{
\item{X}{A 3-dimensional array}

\item{dim}{The dimension along which to compute the mean}
}
\description{
Compute the mean along a cube dimension
}
