% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cube_sum}
\alias{cube_sum}
\title{Compute the sum along a cube dimension}
\usage{
cube_sum(X, dim)
}
\arguments{
\item{X}{A 3-dimensional array}

\item{dim}{The dimension along which to compute the sum}
}
\description{
Compute the sum along a cube dimension
}
