% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{index_to_subscript}
\alias{index_to_subscript}
\title{Vector index to matrix subscripts}
\usage{
index_to_subscript(index, m)
}
\arguments{
\item{index}{a vector of indices}

\item{m}{a matrix (only its size is important)}
}
\description{
Get matrix subscripts from corresponding vector indices (both start from 0).
This is a utility function using Armadillo's ind2sub function.
}
