% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{momscaleBVS}
\alias{momscaleBVS}
\title{A Modular & Multiscale Bayesian Variable Selection model}
\usage{
momscaleBVS(y, Xall, starting, mcmc, gg, module_prior_par,
  binary = FALSE)
}
\arguments{
\item{y}{vector of responses}

\item{Xall}{A list of length K, where each component is a design matrix at a different resolution.}

\item{starting}{A list of length K with the starting configurations. Useful to restart MCMC from good locations}

\item{gg}{g-prior parameter}

\item{module_prior_par}{A vector with K components.  For model M, p(M) is prop. to exp(k * p) where p is the 
number of included variables, and k is model_prior_par}

\item{binary}{Are responses binary (0,1)?}

\item{mcmc_in}{number of Markov chain Monte Carlo iterations}
}
\description{
A Modular & Multiscale Bayesian Variable Selection model
}
