% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvtruncnormal}
\alias{mvtruncnormal}
\title{Sample from Truncated Normal using Botev (2017)}
\usage{
mvtruncnormal(mean, l_in, u_in, Sig, n)
}
\arguments{
\item{mean}{A p-dimensional mean vector}

\item{l_in}{A p-dimensional vector of lower truncation limits (can be -Inf)}

\item{u_in}{A p-dimensional vector of upper truncation limits (can be Inf)}

\item{Sig}{A (p,p) covariance matrix.}

\item{n}{number of samples to extract}
}
\description{
Sample from Truncated Normal using Botev (2017)
}
