% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvtruncnormal_eye1}
\alias{mvtruncnormal_eye1}
\title{Sample from Truncated and shifted Normal with Identity covariance}
\usage{
mvtruncnormal_eye1(mean, l_in, u_in)
}
\arguments{
\item{mean}{A p-dimensional mean vector}

\item{l_in}{A p-dimensional vector of lower truncation limits (can be -Inf)}

\item{u_in}{A p-dimensional vector of upper truncation limits (can be Inf)}
}
\description{
Sample from Truncated and shifted Normal with Identity covariance
}
