% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.r
\name{sofr}
\alias{sofr}
\title{Bayesian Modular & Multiscale Scalar-on-function regression}
\usage{
sofr(y, X, K, mcmc, burn = 0, lambda = 1, family = "gaussian")
}
\arguments{
\item{y}{Vector of outcomes}

\item{X}{Matrix of dimension (n, p)}

\item{K}{Number of modular steps}

\item{mcmc}{Number of Markov chain-Monte Carlo iterations}

\item{burn}{Number of MCMC iterations to discard}

\item{family}{Assumes the outcome is binary with values in (0,1) if set to anything other than "gaussian"}
}
\description{
This function applies a Bayesian Modular & Multiscale regression model to
a scalar response, using functional predictors recorded on a common, regular grid.
}
