% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.r
\name{sofr2}
\alias{sofr2}
\title{Bayesian Modular & Multiscale Scalar-on-function regression}
\usage{
sofr2(y, X, maxk, mcmc, burn = 0, lambda = 1)
}
\arguments{
\item{y}{Vector of outcomes}

\item{X}{Matrix of dimension (n, p)}

\item{mcmc}{Number of Markov chain-Monte Carlo iterations}

\item{burn}{Number of MCMC iterations to discard}

\item{K}{Maximum number of modular steps.}
}
\description{
This function applies a Bayesian Modular & Multiscale regression model to
a scalar response, using functional predictors recorded on a common, regular grid.
The overall model is assumed gaussian.
}
