% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soi_binary}
\alias{soi_binary}
\title{Scalar-on-image Binary regression}
\usage{
soi_binary(y, X, centers, mask_forbid, lambda_centers, lambda_ridge, mcmc,
  burn, radius = 2L, start_movinglev = 0L, partnum = 0L,
  save = TRUE, save_splitmask = TRUE, fixsigma = FALSE)
}
\arguments{
\item{y}{A numeric vector of length n with (0,1) entries}

\item{X}{An array of dimension (p1, p2, n)}

\item{centers}{A list with K elements, each of which is a 2-column matrix. 
Element j of this list will be a matrix listing the (additional) centers to be used 
to split the image space into regions. This parameter only corresponds to the starting value
for MCMC. Only K will remain}

\item{mask_forbid}{A matrix of dimension (p1, p2). Element (i,j) can be set to 0 if the corresponding location 
cannot be used as center, 1 if it can. Typically this should be a matrix of ones.}

\item{lambda_centers}{A scalar parameter controlling the prior probability on the number of centers}

\item{lambda_ridge}{A scalar parameter corresponding to a ridge parameter for each linear regression module}

\item{mcmc}{Number of Markov chain-Monte Carlo iterations}

\item{burn}{Number of MCMC iterations to discard}

\item{radius}{A scalar parameter controlling the dimension of the jumps for the ``move" proposals 
of the centers in MCMC}

\item{start_movinglev}{A scalar smaller or equal to K indicating which}

\item{max_stages}{A scalar controlling the maximum number of stages. Currently fixed at K}
}
\description{
This function applies a Bayesian Modular & Multiscale regression model to
a binary response, using image predictors.
}
