% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.r
\name{soir}
\alias{soir}
\title{Bayesian Modular & Multiscale Scalar-on-image regression}
\usage{
soir(y, X, K, mcmc, burn = 0, family = "gaussian")
}
\arguments{
\item{y}{Vector of outcomes}

\item{X}{Array of dimension (p1, p2, n)}

\item{K}{Number of modular steps}

\item{mcmc}{Number of Markov chain-Monte Carlo iterations}

\item{burn}{Number of MCMC iterations to discard}

\item{family}{Assumes the outcome is binary with values in (0,1) if set to anything other than "gaussian"}
}
\description{
This function applies a Bayesian Modular & Multiscale regression model to
a scalar response, using image predictors.
}
